requirements.txt
pytest-playwright
pytest-xdist
python-dotenv
